Resolved CID53144 (Uninitialized scalar variable) Correct change for https://github.com/AscEmu/AscEmu/commit/771812f7eff6b101157d78bc80fa759f05aa4896 (AIInterface)

+++ src/world/AIInterface.cpp
@@ -638,8 +638,7 @@ void AIInterface::_UpdateCombat(uint32 p_time)
             }
             */
 
-            if (m_canFlee && !m_hasFleed
-                && ((m_Unit->GetHealth() / m_Unit->GetMaxHealth()) < m_FleeHealth))
+            if (m_canFlee && !m_hasFleed && ((static_cast<float>(m_Unit->GetHealth()) / static_cast<float>(m_Unit->GetMaxHealth()) < m_FleeHealth)))
                 agent = AGENT_FLEE;
             else if (m_canCallForHelp
                      && !m_hasCalledForHelp
